requirements.txt
python-dotenv
langchain
langchain-community
langchain-core
langchain-openai
langchain-text-splitters
langgraph
pydantic
faiss-cpu
pypdf
ipykernel
